+++ src/yt_config_output.h
@@ -54,11 +54,11 @@ bool yt_config_compose_genesis_prompt(const uint8_t *current_value,
 bool yt_config_compose_local_beep(size_t initial_column,
     struct yt_config_output_result *result);
 bool yt_config_genesis_valid(float threshold);
-bool yt_config_compose_hq_prompt(float current_hq, float upper_bound,
-    size_t initial_column, struct yt_config_output_result *result);
+bool yt_config_compose_hq_prompt(float current_hq, uint16_t upper_bound,
+	size_t initial_column, struct yt_config_output_result *result);
 bool yt_config_compose_hq_diagnostic(enum yt_config_hq_diagnostic diagnostic,
     size_t initial_column, struct yt_config_output_result *result);
-bool yt_config_hq_in_range(float candidate, float upper_bound);
+bool yt_config_hq_in_range(float candidate, uint16_t upper_bound);
 bool yt_config_compose_scoreboard_prompt(size_t initial_column,
     struct yt_config_output_result *result);
 bool yt_config_compose_scoreboard_too_long(size_t initial_column,
